@@ -4,23 +4,25 @@
     "keywords": ["framework", "laravel"],
     "license": "MIT",
     "type": "project",
+
     "require": {
         "php": ">=5.6.4",
-        "laravel/framework": "5.3.*",
         "barryvdh/laravel-ide-helper": "^2.2",
-        "doctrine/dbal": "^2.5",
-        "predis/predis": "^1.1",
         "cboden/ratchet": "^0.3.5",
-        "zizaco/entrust": "^1.7",
+        "doctrine/dbal": "^2.5",
+        "guzzlehttp/guzzle": "^6.2",
+        "laravel/framework": "5.4.*",
+        "laravel/tinker": "~1.0",
         "mews/captcha": "^2.1",
-        "barryvdh/laravel-debugbar": "^2.3",
-        "guzzlehttp/guzzle": "*",
-        "mews/purifier": "^2.0"
+        "mews/purifier": "^2.0",
+        "predis/predis": "^1.1",
+        "zizaco/entrust": "^1.7",
+        "symfony/psr-http-message-bridge": "^1.0"
     },
     "require-dev": {
         "fzaninotto/faker": "~1.4",
         "mockery/mockery": "0.9.*",
-        "phpunit/phpunit": "~5.0",
+        "phpunit/phpunit": "~5.7",
         "symfony/css-selector": "3.1.*",
         "symfony/dom-crawler": "3.1.*"
     },
@@ -29,17 +31,16 @@
             "database"
         ],
         "psr-4": {
-            "App\\": "app/",
-            "Admin\\": "admin/"
+            "App\\": "app/"
         },
       "files":[
-        "helpers.php"
+        "app/helpers.php"
       ]
     },
     "autoload-dev": {
-        "classmap": [
-            "tests/TestCase.php"
-        ]
+        "psr-4": {
+            "Tests\\": "tests/"
+        }
     },
     "scripts": {
         "post-root-package-install": [
@@ -58,6 +59,5 @@
         ]
     },
     "config": {
-        "preferred-install": "dist"
     }
 }
