@@ -2,7 +2,8 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\User;
+use App\Models\User;
+use App\Service\UserService;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 
@@ -12,7 +13,8 @@ class UserController extends Controller
 {
 
     /** @var  Model $modelClass */
-    protected $modelClass = \App\User::class;
+    protected $modelClass = User::class;
+    protected $service;
     protected $trans;
     protected $name;
     protected $view;
@@ -20,9 +22,10 @@ class UserController extends Controller
     /**
      * UserController constructor.
      */
-    public function __construct()
+    public function __construct(UserService $service)
     {
         $this->trans = trans("model.{$this -> modelClass}");
+        $this->service = $service;
         $this->name = $this->trans['name'];
         $this->view = new \View();
 
@@ -30,9 +33,21 @@ class UserController extends Controller
 //        $this->view->offsetSet('trans',$this->trans);
     }
 
+
     public function index()
     {
-        return view("admin::{$this->name}/index",['trans'=>$this->trans]);
+        return view("admin::{$this->name}/index", ['trans' => $this->trans]);
+    }
+
+    public function create()
+    {
+        $model = new User();
+        return view("admin::{$this->name}/edit", ['model' => $model, 'trans' => $this->trans]);
+    }
+
+    public function store()
+    {
+        $this->service->create();
     }
 
     public function destroy($id)
@@ -43,6 +58,29 @@ class UserController extends Controller
 
     public function edit(User $user)
     {
-        return view("admin::{$this->name}/edit",['model' => $user,'trans'=>$this->trans]);
+        return view("admin::{$this->name}/edit", ['model' => $user, 'trans' => $this->trans]);
+    }
+
+    public function update(Request $request, User $user)
+    {
+
+        $this->validate($request, [
+            'name' => 'required|max:255',
+            'email' => 'required|email|max:255',
+        ]);
+        $params = $request->all();
+        $user->name = data_get($params, 'name');
+        $user->email = data_get($params, 'email');
+        $user->api_token = data_get($params, 'api_token');
+        if ($user->save()) {
+            return redirect(route('admin.user.index'));
+        }
+        return redirect()->back()->withInput()->withErrors(trans('common.edit.failed'));
+
+    }
+
+    public function show(User $user)
+    {
+
     }
 }
