@@ -8,7 +8,6 @@ use Illuminate\Routing\Controller as BaseController;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Arr;
-use Illuminate\Validation\ValidationException;
 
 class Controller extends BaseController
 {
@@ -30,7 +29,7 @@ class Controller extends BaseController
                 'data' => $data,
             ],$statusCodes[$status]);
         }else{
-            return view($status, ['message' => $message]);
+            return view($status, ['message' => $message, 'return_url'=> \Request::get('_return_url')]);
         }
 
     }
