@@ -1,32 +1,35 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: shellus
- * Date: 2017/1/17
- * Time: 10:33
- */
-
-namespace App\Service;
+<?php namespace App\Service;
 
 
-use App\User;
+use App\Models\User;
 use Illuminate\Support\Str;
+use Illuminate\Http\UploadedFile;
+
 
 class UserService
 {
+    protected $fileService;
+
+    public function __construct(FileService $fileService)
+    {
+        $this->fileService = $fileService;
+    }
+
     /**
      * @param $data
      * @return User
      */
-    public static function create($data)
+    public function create($data)
     {
-        return User::create([
+        /** @var User $user */
+        $user = User::create([
             'name' => $data['name'],
             'email' => $data['email'],
             'password' => bcrypt($data['password']),
             'register_ip' => \Request::ip(),
             'api_token' => Str::random(60),
         ]);
+        return $user;
     }
 
     /**
@@ -34,12 +37,32 @@ class UserService
      * @param array $values
      * @return User
      */
-    public static function firstOrCreate($attributes, $values = array())
+    public function firstOrCreate($attributes, $values = array())
     {
         $user = User::where($attributes)->first();
-        if(!$user){
+        if (!$user) {
             $user = self::create($values);
         }
         return $user;
     }
+
+    /**
+     * @param User $user
+     * @param UploadedFile $file
+     */
+    public function updateUserAvatarByUploadedFile(User $user, UploadedFile $file)
+    {
+        $save_path = 'avatar/user_' . $user->getKey();
+        $file_model = $this->fileService->createFormUploadFile($file, $save_path);
+
+
+        // 如果用户当前已有头像则删除先
+        if ($user->avatar()->exists()) {
+            $this->fileService->delete($user->avatar);
+        }
+
+        $user->avatar()->associate($file_model);
+        $user->save();
+    }
+
 }
