@@ -147,7 +147,6 @@ return [
     */
 
     'providers' => [
-
         /*
          * Laravel Framework Service Providers...
          */
@@ -177,6 +176,7 @@ return [
         /*
          * Package Service Providers...
          */
+        Laravel\Tinker\TinkerServiceProvider::class,
 
         // 代码提示
         Barryvdh\LaravelIdeHelper\IdeHelperServiceProvider::class,
@@ -187,8 +187,6 @@ return [
         // 验证码
         Mews\Captcha\CaptchaServiceProvider::class,
 
-        // debugBar
-        Barryvdh\Debugbar\ServiceProvider::class,
 
         // HTML过滤
         Mews\Purifier\PurifierServiceProvider::class,
@@ -218,6 +216,7 @@ return [
     */
 
     'aliases' => [
+        'LogicLog' => App\Facades\LogicLog::class,
 
         'Purifier' => Mews\Purifier\Facades\Purifier::class,
 
@@ -226,11 +225,13 @@ return [
         'Captcha' => Mews\Captcha\Facades\Captcha::class,
 
         'Debugbar' => Barryvdh\Debugbar\Facade::class,
-        
+
         'App' => Illuminate\Support\Facades\App::class,
         'Artisan' => Illuminate\Support\Facades\Artisan::class,
         'Auth' => Illuminate\Support\Facades\Auth::class,
         'Blade' => Illuminate\Support\Facades\Blade::class,
+        'Broadcast' => Illuminate\Support\Facades\Broadcast::class,
+        'Bus' => Illuminate\Support\Facades\Bus::class,
         'Cache' => Illuminate\Support\Facades\Cache::class,
         'Config' => Illuminate\Support\Facades\Config::class,
         'Cookie' => Illuminate\Support\Facades\Cookie::class,
@@ -248,7 +249,7 @@ return [
         'Password' => Illuminate\Support\Facades\Password::class,
         'Queue' => Illuminate\Support\Facades\Queue::class,
         'Redirect' => Illuminate\Support\Facades\Redirect::class,
-        'PRedis' => Illuminate\Support\Facades\Redis::class,
+        'Redis' => Illuminate\Support\Facades\Redis::class,
         'Request' => Illuminate\Support\Facades\Request::class,
         'Response' => Illuminate\Support\Facades\Response::class,
         'Route' => Illuminate\Support\Facades\Route::class,
