@@ -1,7 +1,7 @@
 
 /**
  * First we will load all of this project's JavaScript dependencies which
- * include Vue and Vue Resource. This gives a great starting point for
+ * includes Vue and other libraries. It is a great starting point when
  * building robust, powerful web applications using Vue and Laravel.
  */
 
@@ -14,8 +14,6 @@ require('./bootstrap');
  */
 
 Vue.component('example', require('./components/Example.vue'));
-Vue.component('searchbox', require('./components/SearchBox.vue'));
-
 
 const app = new Vue({
     el: '#app'
