@@ -5,13 +5,13 @@
     <div class="row">
         <div class="col-md-8 col-md-offset-2">
             <div class="panel panel-default">
-                <div class="panel-heading">测试</div>
+                <div class="panel-heading">-</div>
 
                 <div class="panel-body">
-                    测试页面
+                    it's work!
                 </div>
                 <div class="panel-footer">
-                    asd
+                    ---
                 </div>
             </div>
         </div>
