@@ -1,18 +1,7 @@
 <?php
-
-use Illuminate\Foundation\Inspiring;
-
-/*
-|--------------------------------------------------------------------------
-| Console Routes
-|--------------------------------------------------------------------------
-|
-| This file is where you may define all of your Closure based console
-| commands. Each Closure is bound to a command instance allowing a
-| simple approach to interacting with each command's IO methods.
-|
-*/
-
-Artisan::command('inspire', function () {
-    $this->comment(Inspiring::quote());
-});
+/**
+ * Created by PhpStorm.
+ * User: shellus
+ * Date: 2017/4/12
+ * Time: 16:17
+ */
